Finalize header file for spirit protocol
This commits initialize the Headers file forr the class used in the awl translator

+++ spirit-sensor-gateway/spirit-protocol-translation/SpiritProtocol.h
@@ -19,7 +19,7 @@
 
 
 struct SensorTrack {
-    uint64_t id;
+    uint16_t id;
     uint16_t distance;
     int16_t speed;
     int16_t acceleration;
